(169246956) Refactor sign up endpoint - modified sign up controller - modified admin set up - modified middleware functions that validate user on sign up - modified unit tests associated with sign up endpoint - modified all mock data and got rid of unnecesary fields i.e expertise, bio, occupation and address [finishes #169246956]
