--- server/src/v1/tests/MockData/user.js
@@ -1,132 +1,0 @@
-import '../../../../../env';
-
-const User = {
-  user00: {
-    firstName: 'Cindy',
-    lastName: 'Bar',
-    email: 'cindy@bar.com',
-    password: '123456a',
-    address: '4 De Wet Terraces, Godwood, Capetown',
-    occupation: 'Software Engineer',
-    bio: 'I am a full-stack Web Application Developer and Software Developer. I have a Bachelor of Science in Computer Science from University of Capetown, and my primary focus and inspiration for my studies is Web Development. In my free time, I study human computer interface and the psychology of human computer interaction. I am both driven and self-motivated, and I am constantly experimenting with new technologies and techniques. I am very passionate about Web Development',
-    expertise: 'Javascript, React, Angular, Python',
-  },
-
-  user01: {
-    firstName: '',
-    lastName: 'Bar',
-    email: 'cindy@bar.com',
-    password: '123456a',
-    address: '4 De Wet Terraces, Godwood, Capetown',
-    occupation: 'Software Engineer',
-    bio: 'I am a full-stack Web Application Developer and Software Developer. I have a Bachelor of Science in Computer Science from University of Capetown, and my primary focus and inspiration for my studies is Web Development. In my free time, I study human computer interface and the psychology of human computer interaction. I am both driven and self-motivated, and I am constantly experimenting with new technologies and techniques. I am very passionate about Web Development',
-    expertise: 'Javascript, React, Angular, Python',
-  },
-
-  user02: {
-    firstName: 'Cindy',
-    lastName: '',
-    email: 'cindy@bar.com',
-    password: '123456a',
-    address: '4 De Wet Terraces, Godwood, Capetown',
-    occupation: 'Software Engineer',
-    bio: 'I am a full-stack Web Application Developer and Software Developer. I have a Bachelor of Science in Computer Science from University of Capetown, and my primary focus and inspiration for my studies is Web Development. In my free time, I study human computer interface and the psychology of human computer interaction. I am both driven and self-motivated, and I am constantly experimenting with new technologies and techniques. I am very passionate about Web Development',
-    expertise: 'Javascript, React, Angular, Python',
-  },
-
-  user03: {
-    firstName: 'Cindy',
-    lastName: 'Bar',
-    email: 'cindybarcom',
-    password: '123456a',
-    address: '4 De Wet Terraces, Godwood, Capetown',
-    occupation: 'Software Engineer',
-    bio: 'I am a full-stack Web Application Developer and Software Developer. I have a Bachelor of Science in Computer Science from University of Capetown, and my primary focus and inspiration for my studies is Web Development. In my free time, I study human computer interface and the psychology of human computer interaction. I am both driven and self-motivated, and I am constantly experimenting with new technologies and techniques. I am very passionate about Web Development',
-    expertise: 'Javascript, React, Angular, Python',
-  },
-
-  user04: {
-    firstName: 'Cindy',
-    lastName: 'Bar',
-    email: 'cindy@bar.com',
-    password: '2005',
-    address: '4 De Wet Terraces, Godwood, Capetown',
-    occupation: 'Software Engineer',
-    bio: 'I am a full-stack Web Application Developer and Software Developer. I have a Bachelor of Science in Computer Science from University of Capetown, and my primary focus and inspiration for my studies is Web Development. In my free time, I study human computer interface and the psychology of human computer interaction. I am both driven and self-motivated, and I am constantly experimenting with new technologies and techniques. I am very passionate about Web Development',
-    expertise: 'Javascript, React, Angular, Python',
-  },
-
-  user05: {
-    firstName: 'Cindy',
-    lastName: 'Bar',
-    email: 'cindy@bar.com',
-    password: '123456a',
-    address: ' ',
-    occupation: 'Software Engineer',
-    bio: 'I am a full-stack Web Application Developer and Software Developer. I have a Bachelor of Science in Computer Science from University of Capetown, and my primary focus and inspiration for my studies is Web Development. In my free time, I study human computer interface and the psychology of human computer interaction. I am both driven and self-motivated, and I am constantly experimenting with new technologies and techniques. I am very passionate about Web Development',
-    expertise: 'Javascript, React, Angular, Python',
-  },
-
-  user06: {
-    firstName: 'Cindy',
-    lastName: 'Bar',
-    email: 'cindy@bar.com',
-    password: '123456a',
-    address: '4 De Wet Terraces, Godwood, Capetown',
-    occupation: '',
-    bio: 'I am a full-stack Web Application Developer and Software Developer. I have a Bachelor of Science in Computer Science from University of Capetown, and my primary focus and inspiration for my studies is Web Development. In my free time, I study human computer interface and the psychology of human computer interaction. I am both driven and self-motivated, and I am constantly experimenting with new technologies and techniques. I am very passionate about Web Development',
-    expertise: 'Javascript, React, Angular, Python',
-  },
-
-  user07: {
-    firstName: 'Cindy',
-    lastName: 'Bar',
-    email: 'cindy@bar.com',
-    password: '123456a',
-    address: '4 De Wet Terraces, Godwood, Capetown',
-    occupation: 'Software Engineer',
-    bio: '',
-    expertise: 'Javascript, React, Angular, Python',
-  },
-
-  user08: {
-    firstName: 'Cindy',
-    lastName: 'Bar',
-    email: 'cindy@bar.com',
-    password: '123456a',
-    address: '4 De Wet Terraces, Godwood, Capetown',
-    occupation: 'Software Engineer',
-    bio: 'I am a full-stack Web Application Developer and Software Developer. I have a Bachelor of Science in Computer Science from University of Capetown, and my primary focus and inspiration for my studies is Web Development. In my free time, I study human computer interface and the psychology of human computer interaction. I am both driven and self-motivated, and I am constantly experimenting with new technologies and techniques. I am very passionate about Web Development',
-    expertise: '',
-  },
-
-  admin: {
-    firstName: 'admin',
-    lastName: 'admin',
-    email: process.env.ADMIN_EMAIL,
-    password: process.env.ADMIN_PASSWORD,
-    address: '4 De Wet Terraces, Godwood, Capetown',
-    occupation: 'Administrator',
-    bio: 'I am a full-stack Web Application Developer and Software Developer. I have a Bachelor of Science in Computer Science from University of Capetown, and my primary focus and inspiration for my studies is Web Development. In my free time, I study human computer interface and the psychology of human computer interaction. I am both driven and self-motivated, and I am constantly experimenting with new technologies and techniques. I am very passionate about Web Development',
-    expertise: 'Javascript, React, Angular, Python',
-  },
-
-  session00: {
-    mentorId: 70000,
-    questions: `Is the earth flat?
-    How many times in a yea do we have solar ecclipse`,
-  },
-
-  user18: {
-    firstName: 'Foo',
-    lastName: 'Bar',
-    email: 'foo@bar.com',
-    password: '123456a',
-    address: '4 De Wet Terraces, Godwood, Capetown',
-    occupation: 'Software Engineer',
-    bio: 'I am a full-stack Web Application Developer and Software Developer. I have a Bachelor of Science in Computer Science from University of Capetown, and my primary focus and inspiration for my studies is Web Development. In my free time, I study human computer interface and the psychology of human computer interaction. I am both driven and self-motivated, and I am constantly experimenting with new technologies and techniques. I am very passionate about Web Development',
-    expertise: 'Javascript, React, Angular, Python',
-  },
-};
-
-export default User;
